+++ audionimbus/src/effect/path.rs
@@ -162,7 +162,7 @@ use crate::simulation::{SimulationOutputs, Simulator, Source};
 ///     AudioBufferSettings::with_num_channels(NUM_CHANNELS),
 /// )?;
 ///
-/// let simulation_outputs = source.get_outputs(SimulationFlags::PATHING);
+/// let simulation_outputs = source.get_outputs(SimulationFlags::PATHING)?;
 /// let path_effect_params = simulation_outputs.pathing();
 /// let _ = path_effect.apply(&path_effect_params, &input_buffer, &output_buffer);
 /// # Ok::<(), Box<dyn std::error::Error>>(())